pyproject.toml
[project]
name = "utmn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot>=21.7",
    "requests>=2.32.3",
]
